Restructure hs.osascript XPC helper into a class, use @main, and a proper Logger logging infrastructure
